chore(desktop): keep diagnostics changes within repo lint budgets
Problem: the diagnostics integration pushed two existing frontend files past their ratcheted file-size budgets, and the local build left a numeric Wails version stamp behind.

Fix: extract the Virtuoso row/list/footer pieces and project-tree diagnostic emitter into focused modules, preserve the geometry fixture with readable mock data, restore the source Wails version, and baseline the one-line bridge fixture growth required by the browser contract.

Verification: pnpm typecheck; go run ./tools/repolint; git diff --check.
