+++ desktop/frontend/src/components/Transcript.tsx
@@ -1,13 +1,13 @@
-import { forwardRef, lazy, memo, Suspense, type CSSProperties, type KeyboardEvent as ReactKeyboardEvent, type PointerEvent as ReactPointerEvent, type ReactNode, type TouchEvent as ReactTouchEvent, useCallback, useContext, useEffect, useLayoutEffect, useMemo, useRef, useState, useSyncExternalStore, type WheelEvent as ReactWheelEvent } from "react";
-import { Virtuoso, type Components, type ItemProps, type ListItem, type ListProps } from "react-virtuoso";
+import { lazy, Suspense, type CSSProperties, type KeyboardEvent as ReactKeyboardEvent, type PointerEvent as ReactPointerEvent, type ReactNode, type TouchEvent as ReactTouchEvent, useCallback, useEffect, useLayoutEffect, useMemo, useRef, useState, useSyncExternalStore, type WheelEvent as ReactWheelEvent } from "react";
+import { Virtuoso, type ListItem } from "react-virtuoso";
 import type { ControllerLiveStore, Item, LiveStream } from "../lib/useController";
 import type { CheckpointMeta, WireCompletionSummary } from "../lib/types";
 import type { InvocationMetadataMap } from "../lib/invocationDisplay";
 import { useT } from "../lib/i18n";
-import { AssistantMessage, InvocationMetadataContext, TurnActions, UserMessage } from "./Message";
+import { InvocationMetadataContext, TurnActions, UserMessage } from "./Message";
 import { ToolCard } from "./ToolCard";
 import { ExtensionCard } from "./ExtensionCard";
-import { ArrowDown, Loader2, RotateCcw } from "lucide-react";
+import { ArrowDown, Loader2 } from "lucide-react";
 import { Welcome } from "./Welcome";
 import { ReadOnlyBatch } from "./ReadOnlyBatch";
 import { ToolGroup } from "./ToolGroup";
@@ -24,7 +24,6 @@ import {
   foldMapWithReasoningOpen,
   foldMapWithToggle,
   foldSegmentStates,
-  historyEntryIdForRow,
   reconcileFoldEntries,
   splitTranscriptLiveRows,
   EMPTY_FOLDS,
@@ -36,10 +35,8 @@ import {
   type TranscriptRow,
   transcriptRowMeasurementVersion,
 } from "../lib/transcriptRows";
-import { getTranscriptStore } from "../lib/transcriptStore";
 import { createTranscriptMeasuredSizes, type TranscriptSynthesizedSizes } from "../lib/transcriptMeasuredSizes";
 import {
-  estimateTranscriptRowGeometry,
   transcriptRowLayoutVariant,
   type TranscriptEstimateSource,
   type TranscriptGeometryEnvironment,
@@ -51,12 +48,10 @@ import { acquireMarkdownWorkerClient, releaseMarkdownWorkerClient } from "../lib
 import { noteTranscriptRecoveryTerminal, noteTranscriptRowCounts } from "../lib/sessionDiagnostics";
 import { useReasoningDisplayMode } from "../lib/reasoningDisplayPreference";
 import { InlineAssistantReasoning } from "./InlineAssistantReasoning";
-import { LiveTurnRegion } from "./LiveTurnRegion";
 import { ProcessFoldHeader } from "./ProcessFoldHeader";
 import { CompactionCard, NoticeCard, PhaseCard, SteerCard } from "./TranscriptCards";
 import { LiveStreamContext } from "./LiveStreamContext";
 import { useTranscriptSelectableRows } from "../lib/useTranscriptSelectableRows";
-import { TranscriptSelectionOverlay } from "./TranscriptSelectionOverlay";
 import { useCreationTranscriptScrollbar } from "../lib/useCreationTranscriptScrollbar";
 import { useTranscriptScrollInteractions } from "../lib/useTranscriptScrollInteractions";
 import { hasTranscriptScrollableRange, TRANSCRIPT_AT_BOTTOM_THRESHOLD_PX, useTranscriptScrollArbiter } from "../lib/useTranscriptScrollArbiter";
@@ -66,22 +61,23 @@ import { MarkdownImageTabContext } from "./MarkdownImageContext";
 import { recordTranscriptScrollDiagnostic } from "../lib/transcriptScrollProbe";
 import { recordFrontendDiagnostic } from "../lib/frontendDiagnosticBridge";
 import { useTranscriptQuestionJump, useTranscriptQuestions } from "../lib/useTranscriptQuestionNavigation";
+import {
+  LiveAssistantMessage,
+  SHOW_SCROLL_DIAGNOSTICS,
+  TRANSCRIPT_VIRTUOSO_COMPONENTS,
+  TRANSCRIPT_VIRTUOSO_COMPONENTS_WITH_HEADER,
+  type TranscriptVirtuosoContext,
+} from "./TranscriptVirtuosoParts";
 import FrontendDiagnosticsPanelImpl from "./FrontendDiagnosticsPanel";
 
 // NoticeCard lives with the other row cards; keep the historical export path.
 export { NoticeCard } from "./TranscriptCards";
 type OpenTurnAction = { turn: number; menu: "summary" | "rewind" };
 const QUESTION_NAV_MIN_COUNT = 2;
-type AssistantReasoningDisplay = "normal" | "hide";
 const EMPTY_CHECKPOINTS: CheckpointMeta[] = [];
 const EMPTY_INVOCATION_METADATA: InvocationMetadataMap = {};
 const NO_HELD_ROWS: readonly TranscriptRow[] = [];
 const QuestionJumpBar = lazy(() => import("./QuestionJumpBar"));
-const SHOW_SCROLL_DIAGNOSTICS = typeof __BUILD_CHANNEL__ === "undefined"
-  || __BUILD_CHANNEL__ === "test"
-  || __BUILD_CHANNEL__ === "preview"
-  || __BUILD_CHANNEL__ === "canary"
-  || Boolean(import.meta.env?.DEV);
 const SHOW_FRONTEND_DIAGNOSTICS = typeof __BUILD_CHANNEL__ === "undefined"
   || __BUILD_CHANNEL__ === "test"
   || __BUILD_CHANNEL__ === "preview"
@@ -93,210 +89,7 @@ const SHOW_FRONTEND_DIAGNOSTICS = typeof __BUILD_CHANNEL__ === "undefined"
 const FrontendDiagnosticsPanel = SHOW_FRONTEND_DIAGNOSTICS
   ? FrontendDiagnosticsPanelImpl
   : null;
-
-const LiveAssistantMessage = memo(function LiveAssistantMessage({
-  item,
-  defaultExpanded = false,
-  expandWhileStreaming = false,
-  creationMode = false,
-  reasoningDisplay = "normal",
-}: {
-  item: AssistantItem;
-  defaultExpanded?: boolean;
-  expandWhileStreaming?: boolean;
-  creationMode?: boolean;
-  reasoningDisplay?: AssistantReasoningDisplay;
-}) {
-  const live = useContext(LiveStreamContext);
-  const shown = useMemo(
-    () => {
-      const merged =
-        live && live.id === item.id
-          ? {
-              ...item,
-              text: live.text,
-              reasoning: live.reasoning,
-              streaming: true,
-              reasoningComplete: live.reasoningComplete,
-              reasoningDurationMs:
-                live.reasoningStartedAt && live.reasoningCompletedAt && live.reasoningCompletedAt >= live.reasoningStartedAt
-                  ? live.reasoningCompletedAt - live.reasoningStartedAt
-                  : item.reasoningDurationMs,
-            }
-          : item;
-      if (reasoningDisplay === "hide") {
-        return { ...merged, reasoning: "", reasoningComplete: true, reasoningDurationMs: undefined };
-      }
-      return merged;
-    },
-    [item, live?.id, live?.text, live?.reasoning, live?.reasoningComplete, live?.reasoningStartedAt, live?.reasoningCompletedAt, reasoningDisplay],
-  );
-  return (
-    <AssistantMessage
-      item={shown}
-      defaultExpanded={defaultExpanded}
-      expandWhileStreaming={expandWhileStreaming}
-      creationMode={creationMode}
-    />
-  );
-});
 const VIRTUAL_OVERSCAN_ROWS = 8;
-
-type TranscriptVirtuosoContext = {
-  tabId?: string;
-  scrollElement: HTMLDivElement | null;
-  nativeScrollbarDragging: boolean;
-  overlayRevision: string;
-  geometryEnvironment: TranscriptGeometryEnvironment;
-  rowGeometry: {
-    heightEstimates: readonly number[];
-    estimateSources: readonly TranscriptEstimateSource[];
-    rowIndexByKey: ReadonlyMap<string, number>;
-    contentRevision: number;
-  };
-  /** The active turn's in-flow footer region; null when no turn is live. */
-  liveRegion: null | {
-    rows: readonly TranscriptRow[];
-    renderRow: (row: TranscriptRow) => ReactNode;
-    showStatus: boolean;
-    turnStartAt?: number;
-    onPointerDownCapture: (event: ReactPointerEvent<HTMLElement>) => void;
-  };
-  olderHistory: null | {
-    loading: boolean;
-    error?: string;
-    onRetry: () => void;
-  };
-};
-
-const TranscriptVirtuosoItem = forwardRef<HTMLDivElement, ItemProps<TranscriptRow> & { context: TranscriptVirtuosoContext }>(
-  function TranscriptVirtuosoItem({ item, context, children, style, ...props }, ref) {
-    const entryId = historyEntryIdForRow(item);
-    useEffect(() => {
-      if (entryId) getTranscriptStore().requestEntryFullContent(context.tabId, entryId);
-    }, [context.tabId, entryId]);
-    const knownSize = Number.parseFloat(String(props["data-known-size"] ?? ""));
-    const rowIndex = context.rowGeometry.rowIndexByKey.get(String(item.key)) ?? Number.NaN;
-    const estimatedSize = Number.isInteger(rowIndex) && rowIndex >= 0
-      ? context.rowGeometry.heightEstimates[rowIndex]
-      : undefined;
-    const estimateSource = Number.isInteger(rowIndex) && rowIndex >= 0
-      ? context.rowGeometry.estimateSources[rowIndex]
-      : undefined;
-    const layoutVariant = transcriptRowLayoutVariant(item);
-    const staticEstimate = estimateTranscriptRowGeometry(item, context.geometryEnvironment);
-    // Virtuoso can mount a recycled item before the logical-key map has caught
-    // up with its internal data index. Keep a state-aware seed on the DOM in
-    // that narrow window as well; otherwise the manual itemSize freeze falls
-    // through to the stale known-size value and Virtuoso performs its own
-    // upward-compensation write when the row is first visited.
-    const rowEstimate = Number.isFinite(estimatedSize) && (estimatedSize ?? 0) > 0
-      ? estimatedSize
-      : staticEstimate;
-    const diagnosticAttributes = SHOW_SCROLL_DIAGNOSTICS
-      ? {
-          "data-logical-index": Number.isInteger(rowIndex) && rowIndex >= 0 ? rowIndex : undefined,
-          "data-estimated-size": estimatedSize,
-          "data-content-revision": context.rowGeometry.contentRevision,
-          "data-estimate-source": estimateSource,
-        }
-      : {};
-    const frozenStyle = context.nativeScrollbarDragging && Number.isFinite(knownSize) && knownSize > 0
-      ? { ...style, boxSizing: "border-box" as const, height: knownSize, overflow: "hidden" as const }
-      : style;
-    const geometryStyle = Number.isFinite(rowEstimate) && (rowEstimate ?? 0) > 0
-      ? { ...frozenStyle, "--transcript-row-estimate": `${rowEstimate}px` } as CSSProperties
-      : frozenStyle;
-    return (
-      <div
-        {...props}
-        ref={ref}
-        style={geometryStyle}
-        data-row-key={String(item.key)}
-        data-row-kind={item.kind}
-        data-layout-version={transcriptRowMeasurementVersion(item)}
-        data-transcript-layout-variant={layoutVariant}
-        data-transcript-content-width={context.geometryEnvironment.contentWidth}
-        data-transcript-estimate={rowEstimate}
-        data-static-estimate={staticEstimate}
-        {...diagnosticAttributes}
-        className="transcript__row"
-      >
-        {children}
-      </div>
-    );
-  },
-);
-
-const TranscriptVirtuosoList = forwardRef<HTMLDivElement, ListProps & { context: TranscriptVirtuosoContext }>(
-  function TranscriptVirtuosoList({ context, children, ...props }, ref) {
-    return (
-      <div {...props} ref={ref} className="transcript__virtual-sizer">
-        <TranscriptSelectionOverlay
-          tabId={context.tabId ?? ""}
-          scrollElement={context.scrollElement}
-          virtualRevision={context.overlayRevision}
-        />
-        {children}
-      </div>
-    );
-  },
-);
-
-function TranscriptVirtuosoHeader({ context }: { context: TranscriptVirtuosoContext }) {
-  const t = useT();
-  const older = context.olderHistory;
-  if (!older) return null;
-  return (
-    <div className="transcript__header">
-      <div className="transcript__older-status" role={older.error ? "alert" : "status"}>
-        {older.loading ? (
-          <>
-            <Loader2 className="transcript__older-spinner" size={14} aria-hidden="true" />
-            <span>{t("common.loading")}</span>
-          </>
-        ) : (
-          <>
-            <span>{older.error}</span>
-            <button type="button" className="btn btn--small" onClick={older.onRetry}>
-              <RotateCcw size={14} aria-hidden="true" />
-              <span>{t("common.retry")}</span>
-            </button>
-          </>
-        )}
-      </div>
-    </div>
-  );
-}
-
-// The live turn region is the list's in-flow Footer: it scrolls with the
-// transcript but is never part of Virtuoso's measured size tree.
-function TranscriptVirtuosoFooter({ context }: { context: TranscriptVirtuosoContext }) {
-  const live = context.liveRegion;
-  if (!live || (live.rows.length === 0 && !live.showStatus)) return null;
-  return (
-    <LiveTurnRegion
-      rows={live.rows}
-      renderRow={live.renderRow}
-      showStatus={live.showStatus}
-      turnStartAt={live.turnStartAt}
-      tabId={context.tabId}
-      scrollElement={context.scrollElement}
-      onPointerDownCapture={live.onPointerDownCapture}
-    />
-  );
-}
-
-const TRANSCRIPT_VIRTUOSO_COMPONENTS: Components<TranscriptRow, TranscriptVirtuosoContext> = {
-  Item: TranscriptVirtuosoItem,
-  List: TranscriptVirtuosoList,
-  Footer: TranscriptVirtuosoFooter,
-};
-
-const TRANSCRIPT_VIRTUOSO_COMPONENTS_WITH_HEADER: Components<TranscriptRow, TranscriptVirtuosoContext> = {
-  ...TRANSCRIPT_VIRTUOSO_COMPONENTS,
-  Header: TranscriptVirtuosoHeader,
-};
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
