+++ tools/repolint/baseline.json
@@ -10,7 +10,7 @@
     "marker": 0,
     "narrative": 64,
     "struct-state": 115,
-    "test-file-size": 69387
+    "test-file-size": 69271
   },
   "files": {
     "cmd/e2ebench/main.go": {
@@ -32,7 +32,7 @@
     },
     "desktop/app_test.go": {
       "essay": 2,
-      "test-file-size": 10439
+      "test-file-size": 10263
     },
     "desktop/bot_bridge.go": {
       "essay": 3
@@ -130,7 +130,7 @@
       "file-size": 141
     },
     "desktop/frontend/src/components/SettingsPanel.tsx": {
-      "file-size": 7239
+      "file-size": 7160
     },
     "desktop/frontend/src/components/StatusBar.tsx": {
       "file-size": 43
@@ -154,7 +154,7 @@
       "file-size": 180
     },
     "desktop/frontend/src/lib/bridge.ts": {
-      "file-size": 4863
+      "file-size": 4852
     },
     "desktop/frontend/src/lib/crash.ts": {
       "file-size": 225
@@ -163,7 +163,7 @@
       "file-size": 265
     },
     "desktop/frontend/src/lib/types.ts": {
-      "file-size": 1611
+      "file-size": 1604
     },
     "desktop/frontend/src/lib/useController.ts": {
       "file-size": 4035
@@ -258,7 +258,7 @@
     "desktop/settings_app.go": {
       "complexity": 29,
       "essay": 6,
-      "file-size": 2893
+      "file-size": 2828
     },
     "desktop/settings_app_test.go": {
       "test-file-size": 1863
@@ -1536,7 +1536,7 @@
       "file-size": 76
     },
     "internal/config/provider_presets_test.go": {
-      "test-file-size": 174
+      "test-file-size": 126
     },
     "internal/botruntime/runtime.go": {
       "file-size": 38
